Split RateDefinition tests into their own test file
